refactor record token classification
